[#IDENTITY-3639] Secret Activity: extract shared actor-attribution spec helpers
